Use separate parsers for Acrylic (still stubs, though).
