+++ Compiler/Parser/Acrylic/AnnotationParser.cs
@@ -2,6 +2,12 @@
 {
     internal class AnnotationParser : IAnnotationParser
     {
+        private ParserContext parser;
+        public AnnotationParser(ParserContext parser)
+        {
+            this.parser = parser;
+        }
+
         public AnnotationCollection ParseAnnotations(TokenStream tokens)
         {
             throw new System.NotImplementedException();
