+++ Compiler/Parser/Acrylic/ExecutableParser.cs
@@ -5,6 +5,12 @@ namespace Parser.Acrylic
 {
     internal class ExecutableParser : IExecutableParser
     {
+        private ParserContext parser;
+        public ExecutableParser(ParserContext parser)
+        {
+            this.parser = parser;
+        }
+
         public Executable Parse(TokenStream tokens, bool simpleOnly, bool semicolonPresent, Node owner)
         {
             throw new System.NotImplementedException();
