+++ Compiler/Parser/Acrylic/TopLevelParser.cs
@@ -4,6 +4,12 @@ namespace Parser.Acrylic
 {
     internal class TopLevelParser : ITopLevelParser
     {
+        private ParserContext parser;
+        public TopLevelParser(ParserContext parser)
+        {
+            this.parser = parser;
+        }
+
         public TopLevelEntity Parse(TokenStream tokens, TopLevelEntity owner, FileScope fileScope)
         {
             throw new System.NotImplementedException();
